------------------------------------------------------------------------ r4208 | acceptthis | 2013-02-17 05:43:04 +0000 (Sun, 17 Feb 2013) | 2 lines
try to load the replacement skin even if we already tried to load it.
sounds stupid, but the first check checks only for upper/lower, while the second check checks only for base. so even if the first check fails, they might still have a tga, just one that has no colourmapping suppor, so we need the fallback even if we tried it already.
------------------------------------------------------------------------